requirements.txt
# requirements.txt

# Core DL stack
torch>=2.1.0

# Hugging Face ecosystem
transformers>=4.41.0
peft>=0.11.0
sentencepiece>=0.1.99

# Training utilities & metrics
torchmetrics>=1.3.0
tqdm>=4.66.0

# Config & logging
hydra-core>=1.3.2
omegaconf>=2.3.0

# Numerics
numpy>=1.24.0
scipy>=1.10.0
